requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
python-logstash
paho-mqtt